feat: refactor DashboardAside to use handleGoCreateRecordRoute from useDashboard hook
