+++ src/shared/ui/organisms/DashboardAside.tsx
@@ -1,5 +1,4 @@
 import { ReactNode } from "react"
-import { useRouter } from 'next/navigation'
 
 import { HomeIcon } from "../icons/HomeIcon"
 import { CreditCardArrowIcon } from "../icons/CreditCardArrowIcon"
@@ -9,11 +8,8 @@ import { DropdownSelectAccount } from "@/features/Accounts/DropdownSelectAccount
 import { AccountBank } from "@/shared/types/accounts.types"
 import { LinkButton } from "../atoms/LinkButton"
 import { DashboardScreens } from "@/shared/types/dashboard.types"
-import { CREATE_RECORD_ROUTE } from "@/shared/constants/Global.constants"
-import { getAccountCookie } from "@/shared/lib/preferences.lib"
-import { useDashboardStore } from "@/zustand/provider/dashboard-store-provider"
-import { saveAccountApi } from "@/shared/utils/user-info.utils"
 import { Button } from "flowbite-react"
+import { useDashboard } from "@/shared/hooks/useDashboard"
 
 interface DashboardAsideProps {
   children: ReactNode;
@@ -24,21 +20,7 @@ interface DashboardAsideProps {
 }
 
 export const DashboardAside = ({ children, accounts, updateScreen, toggleSelectAccountModal, screen }: DashboardAsideProps) => {
-  const router = useRouter()
-  const selectedAccountId = useDashboardStore(
-      (state) => state.selectedAccount?._id
-    )
-  const handleCreateRecord = async () => {
-    try {
-      const selectedAccountCookie = await getAccountCookie()
-      if (!selectedAccountCookie && selectedAccountId) {
-        await saveAccountApi(selectedAccountId)
-      }
-      router.push(CREATE_RECORD_ROUTE)
-    } catch (error) {
-      console.error('Error creating record:', error)
-    }
-  }
+  const { handleGoCreateRecordRoute } = useDashboard()
   return (
     <aside className="w-72 p-5 flex flex-col gap-4 border-r border-r-gray-600">
       {children}
@@ -46,7 +28,7 @@ export const DashboardAside = ({ children, accounts, updateScreen, toggleSelectA
         <DropdownSelectAccount goAccounts={toggleSelectAccountModal} />
       )}
       <Button
-        onClick={handleCreateRecord}
+        onClick={handleGoCreateRecordRoute}
       >Registrar movimiento</Button>
       <nav className="mt-10 flex flex-col">
         <DashboardAsideLink isSelected={screen === 'overview'} onClickCb={() => updateScreen('overview')}>
